Split GUI normalization workflow
